refactor unit tests into separate modules
